% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U03-weights.R
\name{print.PSweight_estimate}
\alias{print.PSweight_estimate}
\title{Print Method for PSweight_estimate Objects}
\usage{
\method{print}{PSweight_estimate}(x, ...)
}
\arguments{
\item{x}{An object of class "PSweight_estimate".}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print Method for PSweight_estimate Objects
}
