% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A00-surveff.R
\name{print.surveff}
\alias{print.surveff}
\title{Print Method for surveff Objects}
\usage{
\method{print}{surveff}(x, max.len = 6, round.digits = 4, ...)
}
\arguments{
\item{x}{A \code{surveff} object.}

\item{max.len}{Maximum number of rows (time points) to print. Default 6.}

\item{round.digits}{Number of digits for rounding displayed values. Default 4.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for surveff Objects
}
