% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U03-weights.R
\name{summary.PSweight_estimate}
\alias{summary.PSweight_estimate}
\title{Summary Method for PSweight_estimate Objects}
\usage{
\method{summary}{PSweight_estimate}(object, ...)
}
\arguments{
\item{object}{An object of class "PSweight_estimate".}

\item{...}{Additional arguments (currently unused).}
}
\description{
Summary Method for PSweight_estimate Objects
}
