% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U01-data-validation.R
\name{validate_ps_formula}
\alias{validate_ps_formula}
\title{Data Validation Functions for PSsurvival Package}
\usage{
validate_ps_formula(ps_formula)
}
\arguments{
\item{ps_formula}{A formula object for the propensity score model.
Should be of the form `treatment ~ covariates`.}
}
\value{
A list containing:
  \item{formula}{The validated formula object}
  \item{treatment_var}{Character string of the treatment variable name}
}
\description{
Data Validation Functions for PSsurvival Package
}
\keyword{Checks}
\keyword{Formula}
\keyword{Propensity}
\keyword{Score}
\keyword{Validate}
\keyword{and}
\keyword{extracts}
\keyword{formula}
\keyword{from}
\keyword{internal}
\keyword{is}
\keyword{left-hand}
\keyword{name}
\keyword{propensity}
\keyword{score}
\keyword{side.}
\keyword{that}
\keyword{the}
\keyword{treatment}
\keyword{valid}
\keyword{variable}
